+++ components/tree/tree.tsx
@@ -1,26 +1,22 @@
-import {
-    defineComponent,
-    provide,
-    onMounted,
-    watch,
-    ref,
-    VNodeChild,
-} from 'vue';
-import { isFunction, isString, cloneDeep, debounce } from 'lodash-es';
+import { defineComponent, provide, VNodeChild } from 'vue';
+import { isFunction, isString } from 'lodash-es';
 import getPrefixCls from '../_util/getPrefixCls';
 import { useTheme } from '../_theme/useTheme';
 import VirtualList from '../virtual-list/virtualList';
 import TreeNode from './treeNode';
-import { COMPONENT_NAME, CHECK_STRATEGY } from './const';
+import { COMPONENT_NAME } from './const';
 import useData from './useData';
 import useState from './useState';
 import useDrag from './useDrag';
 import useFilter from './useFilter';
+import useExpand from './useExpand';
+import useSelect from './useSelect';
+import useCheck from './useCheck';
+import useCurrentData from './useCurrentData';
 
 import { treeProps, TREE_PROVIDE_KEY } from './props';
 
-import type { InnerTreeOption, TreeNodeKey } from './interface';
-import { concat } from '../_util/utils';
+import type { TreeNodeKey } from './interface';
 
 const prefixCls = getPrefixCls('tree');
 
@@ -48,6 +44,11 @@ export default defineComponent({
     setup(props, { emit, expose }) {
         useTheme();
 
+        const { nodeList, allKeys } = useData({
+            props,
+            emit,
+        });
+
         const {
             currentExpandedKeys,
             updateExpandedKeys,
@@ -56,223 +57,48 @@ export default defineComponent({
             currentSelectedKeys,
             updateSelectedKeys,
             hasSelected,
-        } = useState(props, { emit });
-
-        const { nodeList, transformData } = useData({
-            props,
-        });
+        } = useState({ props, emit });
 
         const { filter, filteredExpandedKeys, filteredKeys, isSearchingRef } =
-            useFilter(props, transformData, nodeList);
+            useFilter(props, allKeys, nodeList);
 
-        watch(
-            transformData,
-            () => {
-                emit('update:nodeList', nodeList);
-            },
-            {
-                immediate: true,
-            },
-        );
-
-        let checkingNode: InnerTreeOption;
-
-        function computeIndeterminate(key: TreeNodeKey) {
-            const node = nodeList.get(key);
-            if (node.hasChildren) {
-                if (node.isChecked.value) {
-                    node.isIndeterminate.value = false;
-                } else {
-                    node.isIndeterminate.value = node.children.some(
-                        (item) =>
-                            item.isChecked.value || item.isIndeterminate.value,
-                    );
-                }
-            } else {
-                node.isIndeterminate.value = false;
-            }
-        }
-
-        watch(currentCheckedKeys, (newKeys, oldKeys) => {
-            // 重置历史节点选择状态
-            oldKeys.forEach((key: TreeNodeKey) => {
-                const node = nodeList.get(key);
-                node.isChecked.value = false;
-            });
-            newKeys.forEach((key: TreeNodeKey) => {
-                const node = nodeList.get(key);
-                node.isChecked.value = true;
-            });
-            if (props.cascade) {
-                // 当选中某个节点时，只需要处理此节点相关上下节点状态
-                if (checkingNode) {
-                    const { indexPath } = checkingNode;
-                    indexPath.slice(0).reverse().forEach(computeIndeterminate);
-                    checkingNode.hasChildren &&
-                        checkingNode.childrenPath.forEach(
-                            (key: TreeNodeKey) => {
-                                const node = nodeList.get(key);
-                                node.isIndeterminate.value = false;
-                            },
-                        );
-                    checkingNode = null;
-                } else {
-                    transformData.value.forEach(computeIndeterminate);
-                }
-            }
+        const { expandNode, expandingNode } = useExpand({
+            allKeys,
+            isSearchingRef,
+            filteredExpandedKeys,
+            nodeList,
+            currentExpandedKeys,
+            updateExpandedKeys,
+            props,
+            emit,
         });
 
-        const currentData = ref<TreeNodeKey[]>([]);
-
-        let expandingNode: InnerTreeOption;
-
-        const _addNode = (
-            nodes: InnerTreeOption[],
-            res: TreeNodeKey[] = [],
-        ) => {
-            nodes.forEach((node) => {
-                res.push(node.value);
-                if (node.hasChildren && node.isExpanded.value) {
-                    _addNode(node.children, res);
-                }
-            });
-        };
-
-        const addNode = (nodes: InnerTreeOption[], index: number) => {
-            const res: TreeNodeKey[] = [];
-            _addNode(nodes, res);
-            const arr1 = currentData.value.slice(0, index);
-            const arr2 = currentData.value.slice(index);
-            concat(arr1, res);
-            concat(arr1, arr2);
-            currentData.value = arr1;
-        };
-
-        const deleteNode = (keys: TreeNodeKey[], index: number) => {
-            let len = 0;
-            keys.forEach((key) => {
-                if (key === currentData.value[index + len]) {
-                    len += 1;
-                }
-            });
-            currentData.value.splice(index, len);
-        };
-
-        const computeCurrentData = () => {
-            const res: TreeNodeKey[] = [];
-            const expandedKeys = isSearchingRef.value
-                ? filteredExpandedKeys.value
-                : currentExpandedKeys.value;
-
-            const keys = isSearchingRef.value
-                ? filteredKeys.value
-                : transformData.value;
-
-            if (expandingNode) {
-                // 展开后
-                if (expandingNode.isExpanded.value) {
-                    const index = currentData.value.indexOf(
-                        expandingNode.value,
-                    );
-                    addNode(expandingNode.children, index + 1);
-                } else {
-                    const index = currentData.value.indexOf(
-                        expandingNode.value,
-                    );
-                    deleteNode(expandingNode.childrenPath, index + 1);
-                }
-                expandingNode = null;
-                return;
-            }
-
-            // 缓存每个节点的展开状态，性能更优
-            keys.forEach((key) => {
-                const node = nodeList.get(key);
-                if (node.hasChildren) {
-                    node.isExpanded.value = expandedKeys.includes(key);
-                }
-                const indexPath = node.indexPath;
-                const len = indexPath.length;
-                let index = 0;
-                let parentExpanded = true;
-                while (index < len - 1) {
-                    const parentNode = nodeList.get(indexPath[index]);
-                    if (!parentNode.isExpanded.value) {
-                        parentExpanded = false;
-                        break;
-                    }
-                    index += 1;
-                }
-                if (parentExpanded) {
-                    res.push(key);
-                }
-            });
-            currentData.value = res;
-        };
-
-        watch(
-            [filteredExpandedKeys, filteredKeys],
-            debounce(() => {
-                if (!isSearchingRef.value) return;
-                computeCurrentData();
-            }, 10),
-        );
-
-        watch(
-            [currentExpandedKeys, transformData],
-            debounce(() => {
-                if (isSearchingRef.value) return;
-                computeCurrentData();
-            }, 10),
-            {
-                immediate: true,
-            },
-        );
-
-        watch([isSearchingRef], () => {
-            computeCurrentData();
+        const { selectNode } = useSelect({
+            nodeList,
+            currentSelectedKeys,
+            updateSelectedKeys,
+            props,
+            emit,
         });
 
-        const expandNode = (val: TreeNodeKey, event: Event) => {
-            if (isSearchingRef.value) {
-                const _value = cloneDeep(filteredExpandedKeys.value);
-                const index = _value.indexOf(val);
-                // 已经展开
-                if (index !== -1) {
-                    _value.splice(index, 1);
-                } else {
-                    _value.push(val);
-                }
-                filteredExpandedKeys.value = _value;
-                return;
-            }
-            const node = nodeList.get(val);
-            expandingNode = node;
-            let values: TreeNodeKey[] = cloneDeep(currentExpandedKeys.value);
-            const index = values.indexOf(val);
-            // 已经展开
-            if (index !== -1) {
-                values.splice(index, 1);
-                // 让动画早点动起来
-                node.isExpanded.value = false;
-            } else {
-                if (props.accordion) {
-                    values = values.filter((item) =>
-                        node.indexPath.includes(item),
-                    );
-                }
-                values.push(val);
-                // 让动画早点动起来
-                node.isExpanded.value = true;
-            }
-            updateExpandedKeys(values);
-            emit('expand', {
-                expandedKeys: values,
-                event,
-                node,
-                expanded: values.includes(val),
-            });
-        };
+        const { checkNode } = useCheck({
+            allKeys,
+            nodeList,
+            currentCheckedKeys,
+            updateCheckedKeys,
+            props,
+            emit,
+        });
+
+        const { currentData } = useCurrentData({
+            isSearchingRef,
+            filteredExpandedKeys,
+            currentExpandedKeys,
+            filteredKeys,
+            allKeys,
+            expandingNode,
+            nodeList,
+        });
 
         const {
             handleDragstart,
@@ -283,145 +109,6 @@ export default defineComponent({
             handleDrop,
             dragOverInfo,
         } = useDrag({ nodeList, emit, expandNode });
-
-        onMounted(() => {
-            if (
-                props.defaultExpandAll &&
-                currentExpandedKeys.value.length === 0
-            ) {
-                updateExpandedKeys(
-                    transformData.value.filter(
-                        (value) => !nodeList.get(value).isLeaf,
-                    ),
-                );
-            }
-        });
-
-        const selectNode = (val: TreeNodeKey, event: Event) => {
-            if (!props.selectable) {
-                return;
-            }
-            const node = nodeList.get(val);
-            const values = cloneDeep(currentSelectedKeys.value);
-            const index = values.indexOf(val);
-            if (props.multiple) {
-                if (index !== -1) {
-                    props.cancelable && values.splice(index, 1);
-                } else {
-                    values.push(val);
-                }
-            } else if (index !== -1) {
-                props.cancelable && values.splice(index, 1);
-            } else {
-                values[0] = val;
-            }
-            updateSelectedKeys(values);
-            emit('select', {
-                selectedKeys: values,
-                event,
-                node,
-                selected: values.includes(val),
-            });
-        };
-
-        function getCheckedKeys(arr: TreeNodeKey[]) {
-            return props.cascade
-                ? arr.filter((key) => {
-                      const node = nodeList.get(key);
-                      if (props.checkStrictly === CHECK_STRATEGY.ALL) {
-                          return true;
-                      }
-                      if (props.checkStrictly === CHECK_STRATEGY.PARENT) {
-                          return (
-                              node.indexPath.filter((path) =>
-                                  arr.includes(path),
-                              ).length === 1
-                          );
-                      }
-                      if (props.checkStrictly === CHECK_STRATEGY.CHILD) {
-                          return node.isLeaf;
-                      }
-                      return true;
-                  })
-                : arr;
-        }
-        function handleChildren(
-            arr: TreeNodeKey[],
-            children: InnerTreeOption[],
-            isAdd: boolean,
-        ) {
-            children &&
-                children.forEach((child) => {
-                    const index = arr.indexOf(child.value);
-                    if (!isAdd) {
-                        if (index !== -1) {
-                            arr.splice(index, 1);
-                        }
-                    } else if (index === -1) {
-                        arr.push(child.value);
-                    }
-                    if (child.children) {
-                        handleChildren(arr, child.children, isAdd);
-                    }
-                });
-        }
-        function handleParent(
-            arr: TreeNodeKey[],
-            indexPath: TreeNodeKey[],
-            isAdd: boolean,
-        ) {
-            let len = indexPath.length - 2;
-            for (len; len >= 0; len--) {
-                const parent = nodeList.get(indexPath[len]);
-                const index = arr.indexOf(parent.value);
-                if (!isAdd) {
-                    if (index !== -1) {
-                        arr.splice(index, 1);
-                    }
-                } else if (index === -1) {
-                    if (
-                        parent.children.every((item) =>
-                            arr.includes(item.value),
-                        )
-                    ) {
-                        arr.push(parent.value);
-                    }
-                }
-            }
-        }
-        const checkNode = (val: TreeNodeKey, event: Event) => {
-            const node = nodeList.get(val);
-            const { isLeaf, children, indexPath } = node;
-            const values = cloneDeep(currentCheckedKeys.value);
-            const index = values.indexOf(val);
-            if (!props.cascade) {
-                if (index !== -1) {
-                    values.splice(index, 1);
-                } else {
-                    values.push(val);
-                }
-            } else if (index !== -1) {
-                values.splice(index, 1);
-                handleParent(values, indexPath, false);
-                if (!isLeaf) {
-                    handleChildren(values, children, false);
-                }
-            } else {
-                values.push(val);
-                handleParent(values, indexPath, true);
-                if (!isLeaf) {
-                    handleChildren(values, children, true);
-                }
-            }
-            checkingNode = node;
-            updateCheckedKeys(values);
-            emit('check', {
-                checkedKeys: getCheckedKeys(values),
-                event,
-                node,
-                checked: values.includes(val),
-            });
-        };
 
         if (expose) {
             expose({
